requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
coverage==7.5.3
cryptography==42.0.8
cycler==0.12.1
Django==5.0.6
django-debug-toolbar==4.4.2
django-flatpickr==2.0.3
django-webpack-loader==3.1.0
djangorestframework==3.15.1
fonttools==4.53.0
idna==3.7
kiwisolver==1.4.5
kombu==5.3.7
matplotlib==3.9.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.6
reportlab==4.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tenacity==8.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
